% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grapple_stat_multi_cpp}
\alias{grapple_stat_multi_cpp}
\title{GRAPPLE stat (multivariable) in C++ with OpenMP}
\usage{
grapple_stat_multi_cpp(RxyList, theta, n_threads = 1L)
}
\arguments{
\item{RxyList}{A (p+1) x (p+1) x m cube, the k-th slice is the
(p+1) x (p+1) matrix for SNP k.}

\item{theta}{A length-p vector of current causal effect estimates.}

\item{n_threads}{Number of OpenMP threads.}
}
\value{
List with var_vec (length m), var_cor (m x p),
  bias_correction (p x p).
}
\description{
GRAPPLE stat (multivariable) in C++ with OpenMP
}
