% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grapple_stat_uni_cpp}
\alias{grapple_stat_uni_cpp}
\title{GRAPPLE stat (univariable) in C++ with OpenMP}
\usage{
grapple_stat_uni_cpp(RxyList, theta, n_threads = 1L)
}
\arguments{
\item{RxyList}{A 2 x 2 x m cube. The k-th slice is the 2 x 2
matrix for SNP k.}

\item{theta}{Scalar causal effect estimate.}

\item{n_threads}{Number of OpenMP threads.}
}
\value{
List with var_vec (length m), var_cor (length m),
  bias_correction (scalar).
}
\description{
GRAPPLE stat (univariable) in C++ with OpenMP
}
