% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRBEE_BBC.R
\name{MRBEE_BBC}
\alias{MRBEE_BBC}
\title{MRBEE with background bias correction (BBC).}
\usage{
MRBEE_BBC(
  by,
  bX,
  byse,
  bXse,
  gcov,
  ldsc,
  cov_RB,
  max.iter = 30,
  max.eps = 1e-04,
  pv.thres = 0.05,
  var.est = "variance",
  FDR = T,
  adjust.method = "Sidak",
  sampling.time = 300,
  sampling.strategy = "subsampling",
  n_threads = 4
)
}
\arguments{
\item{by}{A vector (n x 1) of the GWAS effect size of outcome yielded by Rao-Blackwell Correction.}

\item{bX}{A matrix (n x p) of the GWAS effect sizes of p exposures yielded by Rao-Blackwell Correction.}

\item{byse}{A vector (n x 1) of the GWAS effect size SE of outcome yielded by Rao-Blackwell Correction.}

\item{bXse}{A matrix (n x p) of the GWAS effect size SEs of p exposuresyielded by Rao-Blackwell Correction.}

\item{gcov}{A matrix (p+1 x p+1) of the per-snp genetic covariance matrix of the p exposures and outcome. The last one should be the outcome.}

\item{ldsc}{A vector (n x 1) of the LDSCs of the IVs.}

\item{cov_RB}{A list of m matrices of Rao-Blackwell correction terms, each of dimension (p+1) × (p+1).}

\item{max.iter}{Maximum number of iterations for causal effect estimation. Defaults to 30.}

\item{max.eps}{Tolerance for stopping criteria. Defaults to 1e-4.}

\item{pv.thres}{P-value threshold in pleiotropy detection. Defaults to 0.05.}

\item{var.est}{Method for estimating the variance of residual in pleiotropy test. Can be "robust", "variance", or "ordinal". Defaults is "variance" that estimates the variance of residual using median absolute deviation (MAD).}

\item{FDR}{Logical. Whether to apply the FDR to convert the p-value to q-value. Defaults to TRUE.}

\item{adjust.method}{Method for estimating q-value. Defaults to "Sidak".}

\item{sampling.time}{Number of bootstrap iterations used to estimate the
covariance of the causal effect estimates. Defaults to 300.}

\item{sampling.strategy}{Character string indicating the resampling
strategy, either \code{"bootstrap"} (sampling with replacement) or
\code{"subsampling"} (using 50\% of the instruments without replacement).}

\item{n_threads}{The number of threads. Defaults to 4.}
}
\value{
A list containing the estimated causal effect, its covariance, and pleiotropy
}
\description{
This function estimates the causal effect using a bias-correction estimating equation, combining with background bias correction (BBC).
}
