% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRBEE_UV_BBC.R
\name{MRBEE_UV_BBC}
\alias{MRBEE_UV_BBC}
\title{Univariable MRBEE with background bias correction (BBC)}
\usage{
MRBEE_UV_BBC(
  by,
  bx,
  byse,
  bxse,
  cov_RB,
  gcov = diag(2) * 0,
  ldsc = rep(0, length(by)),
  max.iter = 50,
  max.eps = 1e-06,
  sampling.time = 300,
  sampling.strategy = "subsampling",
  n_threads,
  pv.thres = 0.05,
  var.est = "variance",
  FDR = T,
  adjust.method = "Sidak"
)
}
\arguments{
\item{by}{A numeric vector (m × 1) of GWAS effect sizes for the outcome yielded by Rao-Blackwell Correction.}

\item{bx}{A numeric vector (m × 1) of GWAS effect sizes for the exposure yielded by Rao-Blackwell Correction.}

\item{byse}{A numeric vector (m × 1) of standard errors for `by` yielded by Rao-Blackwell Correction.}

\item{bxse}{A numeric vector (m × 1) of standard errors for `bx` yielded by Rao-Blackwell Correction.}

\item{cov_RB}{A list of m matrices of Rao-Blackwell correction terms, each of dimension 2 × 2.}

\item{gcov}{A matrix (2 x 2) of the per-snp genetic covariance matrix of the p exposures and outcome. The last one should be the outcome.}

\item{ldsc}{A vector (n x 1) of the LDSCs of the IVs.}

\item{max.iter}{Maximum number of iterations for updating the causal effect. Default is 30.}

\item{max.eps}{Tolerance for stopping criteria. Default is 1e-4.}

\item{sampling.time}{Number of bootstrap iterations used to estimate the
variance of the causal effect. Default is 300.}

\item{sampling.strategy}{"bootstrap" or "subsampling" (0.5 sample without replacement).}

\item{n_threads}{The number of threads. Defaults to 4.}

\item{pv.thres}{P-value threshold used in pleiotropy detection. Default is 0.05.}

\item{var.est}{Method for estimating the variance of residuals. One of "robust", "variance", or "ordinal". Default is "variance".}

\item{FDR}{Logical. Whether to convert p-values to q-values using FDR adjustment. Default is TRUE.}

\item{adjust.method}{Method used to adjust p-values. Default is "Sidak".}
}
\value{
A list with the following components:
\item{theta}{Estimated causal effect.}
\item{vartheta}{Estimated variance of theta.}
\item{delta}{Estimated pleiotropic residuals (set to zero for valid IVs).}
}
\description{
This function estimates the univariable causal effect using bias-corrected estimating equations with background bias correction (BBC) to improve numerical stability under weak instruments.
}
