% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MRcML_bXest}
\alias{MRcML_bXest}
\title{MRcML_bXest in C++ with OpenMP (multivariable)}
\usage{
MRcML_bXest(ThetaList, bX, e, theta, n_threads = 1L)
}
\arguments{
\item{ThetaList}{(p+1) x (p+1) x m array, as arma::cube}

\item{bX}{m x p matrix of original exposure effects}

\item{e}{length-m vector of residuals}

\item{theta}{length-p vector of causal effects}

\item{n_threads}{number of OpenMP threads}
}
\value{
m x p matrix of bXest
}
\description{
MRcML_bXest in C++ with OpenMP (multivariable)
}
