% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MRcML_UV_bxest}
\alias{MRcML_UV_bxest}
\title{MRcML_UV_bxest in C++ with OpenMP (univariable)}
\usage{
MRcML_UV_bxest(ThetaList, bx, e, theta, n_threads = 1L)
}
\arguments{
\item{ThetaList}{2 x 2 x m array, as arma::cube}

\item{bx}{length-m vector of exposure effects}

\item{e}{length-m vector of residuals}

\item{theta}{scalar causal effect}

\item{n_threads}{number of OpenMP threads}
}
\value{
length-m vector of bxest
}
\description{
MRcML_UV_bxest in C++ with OpenMP (univariable)
}
