% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaoBlackwellCorrect_R.R
\name{RaoBlackwellCorrect_R}
\alias{RaoBlackwellCorrect_R}
\title{Rao-Blackwell Correction for Selected IVs in MVMR}
\usage{
RaoBlackwellCorrect_R(
  BETA_Select,
  SE_Select,
  Rxy,
  eta = 1,
  cutoff,
  B = 1000,
  onlyexposure = T
)
}
\arguments{
\item{BETA_Select}{A numeric matrix of dimension m x p: effect estimates for selected IVs. Outcome should be the last column.}

\item{SE_Select}{A numeric matrix of same dimension: standard errors. Outcome should be the last column.}

\item{Rxy}{A (p+1) x (p+1) numeric matrix: covariance matrix of Z statistics. Outcome should be the last column and row.}

\item{eta}{Standard deviation of noise added to Z (default = 1).}

\item{cutoff}{Chi-square threshold used for selection.}

\item{B}{Number of samples used in rejection sampling (default = 1000).}

\item{onlyexposure}{A indicator of whether considering outcome when selecting IVs. Defaults to \code{TRUE}.}
}
\value{
A list containing:
\describe{
  \item{\code{BETA_RB}}{Corrected effect estimates (m x p).}
  \item{\code{SE_RB}}{Corrected standard errors (m x p).}
}
}
\description{
Performs Rao-Blackwell correction via rejection sampling for each selected IV
based on randomized multivariate chi-square test.
}
