% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaoBlackwellCorrect_UV_R.R
\name{RaoBlackwellCorrect_UV_R}
\alias{RaoBlackwellCorrect_UV_R}
\title{Rao-Blackwell Correction via Resampling for Univariate MR}
\usage{
RaoBlackwellCorrect_UV_R(
  gamma,
  sigma,
  cutoff,
  B = 1000,
  eta = 1,
  n_threads = round(parallel::detectCores()/2)
)
}
\arguments{
\item{gamma}{A numeric vector of effect estimates (e.g. SNP-exposure).}

\item{sigma}{A numeric vector of standard errors, same length as \code{gamma}.}

\item{cutoff}{A numeric threshold on the randomized Z-statistic (i.e., abs(Z + noise) > cutoff).}

\item{B}{Number of Monte Carlo samples to use in rejection sampling (default = 1000).}

\item{eta}{Standard deviation of added noise (default = 1).}

\item{n_threads}{Number of threads to use (default = parallel::detectCores() / 2).}
}
\value{
A data.frame with columns:
\describe{
  \item{\code{BETA_RB}}{Rao-Blackwell corrected effect estimates.}
  \item{\code{SE_RB}}{Estimated standard errors based on resampling.}
}
}
\description{
Performs Rao-Blackwellized correction of SNP-exposure estimates via rejection sampling
using randomized Z-scores. This version wraps a fast C++ implementation.
}
