% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RB_UV_Plink_Joint.R
\name{RB_UV_Plink_Joint}
\alias{RB_UV_Plink_Joint}
\title{Joint Rao-Blackwellized Correction for Univariate MR (Exposure + Outcome)}
\usage{
RB_UV_Plink_Joint(
  GWAS_X,
  GWAS_Y,
  Rxy,
  bedfile,
  p1 = 5e-08,
  p2 = 5e-08,
  r2 = 0.001,
  kb = 1000,
  plink_path = "./plink",
  log_dir = NULL,
  eta = 1,
  B = 1000
)
}
\arguments{
\item{GWAS_X}{Data.frame with columns "SNP", "BETA", "SE" for the exposure.}

\item{GWAS_Y}{Data.frame with columns "SNP", "BETA", "SE" for the outcome, rows must match GWAS_X.}

\item{Rxy}{2x2 correlation/covariance matrix for estimation errors (rows/columns: exposure and outcome).}

\item{bedfile}{Path to PLINK .bed/.bim/.fam prefix.}

\item{p1, p2, r2, kb}{PLINK clumping parameters.}

\item{plink_path}{Path to PLINK binary.}

\item{log_dir}{Temporary directory to store intermediate files.}

\item{eta}{Noise sd for randomized selection Z.}

\item{B}{Number of Monte Carlo samples in rejection sampling.}
}
\value{
A list with two data.frames: GWAS_X_RB and GWAS_Y_RB
}
\description{
Joint Rao-Blackwellized Correction for Univariate MR (Exposure + Outcome)
}
