% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_adj_cov.R
\name{reliability_adj_cov}
\alias{reliability_adj_cov}
\title{Reliability adjustment for Rao-Blackwell correction covariance matrices}
\usage{
reliability_adj_cov(bX, Cov_RB, threshold = 0.5)
}
\arguments{
\item{bX}{A matrix (n x p) of GWAS effect sizes of p exposures.}

\item{Cov_RB}{A list of n matrices of Rao-Blackwell correction terms,
each of dimension (p+1) x (p+1).}

\item{threshold}{Minimum reliability threshold. Defaults to 0.5.}
}
\value{
A list of n adjusted covariance matrices.
}
\description{
For each IV, limits the proportion of estimation error relative to total
variance to prevent unstable estimation caused by extremely weak instruments.
}
